added manage listings

+++ app/Http/Controllers/ListingController.php
@@ -66,6 +66,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store("logos", 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing created successfully!');
@@ -82,6 +84,11 @@ class ListingController extends Controller
     {
         // the dd (dye dump?) does an in browser console.log
         // dd($request->all());
+
+        // Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
 
         $formFields = $request->validate([
             'title' => 'required',
@@ -105,8 +112,19 @@ class ListingController extends Controller
     // Delete Listing
     public function delete(Listing $listing)
     {
+        // Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
 
         return redirect('/')->with('message', 'Listing deleted successfully!');
+    }
+
+    // Manage Listings
+    public function manage()
+    {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
